Renamed package "number" to "numeric" which makes a little more sense and will be used elsewhere in tests in the future
