+++ src/test/java/jnr/ffi/numeric/utils/NumberTestLib.java
@@ -1,4 +1,4 @@
-package jnr.ffi.number.utils;
+package jnr.ffi.numeric.utils;
 
 import jnr.ffi.NativeLong;
 import jnr.ffi.annotations.LongLong;
